+++ src/test/java/net/imagej/ops/thread/RunInterleavedChunker.java
@@ -55,7 +55,7 @@ public class RunInterleavedChunker<A extends RealType<A>> extends
 
 			@Override
 			public void
-				execute(int startIndex, final int stepSize, final int numSteps)
+				execute(long startIndex, final long stepSize, final long numSteps)
 			{
 				final Cursor<A> cursor = input.localizingCursor();
 				final Cursor<A> cursorOut = output.localizingCursor();
